worker: move LF check functions to separate file
